Get zone from PKS master node
PKS vms are created by bosh in the zone bosh is configued to use. It may
not be the same zone that gcloud is defaulted to. In that case, when
configuring external access to the k8s api server, the zone of the
master node must be used.

Also configuring external DNS, for kicks