Convert function .fats/create to riff.toml
By using riff.toml we increase the usability of the functions because
the caller can leverage the default configuration for the function
rather than being required to pass the config via custom means.

One downside of this change is that we are no longer testing the cli
flags to override the default values.

Refs projectriff/riff#1192