Fail remote PRs (#184)
Disabling remote PRs is insufficient as the tests don't run and GitHub
reports the PR checks as passing, because none failed. We need to
actively fail PR from remote repos.

PRs from the same repo are allowed to pass, but are shortcutted to avoid
duplicate builds from the pushed commit.