Replace Minikube with kind
Kind (Kubernetes IN Docker) is a conformant k8s runtime optimized for
testing and CI environments. To run unauthenticated FATS jobs, we
previously used Minikube, which is optimized for developer desktop
environments and was a bit tacky and fragile in CI. (I had major trouble
getting either Minikube 1.4 or K8s 1.14 working in CI without a lot of
hoops, kind just worked).

As a major perk, running with the docker-daemon registry is much faster,
and is only marginally slower than dockerhub (with Minikube runs took
twice as long).

The Minikube scripts are still in this repo, but are now untested and
should be removed once other project have migrated to Kind.