Use distinct repo for each image (#261)
CNB will build images with the same digest given the same buildpacks and
source. This is a good thing, but means that concurrent FATS runs may
collide if they happen to be building images in the same repo.

This change creates a unique repo for each image rather then using tags
within the same repo to distinguish images.