Clarify that FATS install, start and cleanup script are only intended for CI
FATS is designed for bootstrapping CI environments that will be thrown
away once the tests are complete. They make many assumptions about the
operating system, available tools, root access, etc that may not be
appropriate for other environments.

The functions, applications, and macros are general purpose and may be
used anywhere.

Refs #249