Cleanup tests
Avoid DNS caches inside the dev-utils pod when invoking different
deployers with the same name. The DNS TTL is 30s by default, but we're
invoking, deleting the deployer, creating the deployer and invoking
within the window. When that happens the second invoke will try to
connect to a service ip that is no exists.